+++ CHANGELOG.md
@@ -13,11 +13,14 @@ This file tracks changes to this project. It follows the [Keep a Changelog forma
 - `NewFiniteReplayProvider` is no more.
 - `Server.Subscribe` is no more – it never made sense.
 - `Server.Provider` is no more.
+- `NewJoe` is no more.
+- `JoeConfig` is no more.
 
 ### Added
 
 - Because the `ValidReplayProvider` constructor was removed, the fields `ValidReplayProvider.{TTL,AutoIDs}` were added for configuration.
 - Because the `FiniteReplayProvider` constructor was removed, the fields `FiniteReplayProvider.{Count,AutoIDs}` were added for configuration.
+- Because the `Joe` constructor was removed, the fields `Joe.{ReplayProvider,ReplayGCInterval}` were added for configuration.
 
 ### Changed